refactor(apollo): move tests directory
